Added version to dynamic ui layout
* so that you can increase this number and be sure, that it get's relaoded, because sometimes field changes where made, that didn't result in a different hashcode
* always use double in edit parameter fab, because and int can be used as a double but not the other way arround (might not be true, but this will be seen)

+++ lib/devices/generic_device.dart
@@ -434,8 +434,8 @@ class GenericDeviceScreenState extends ConsumerState<GenericDeviceScreen> {
     } else if (valueModel.currentValue is int) {
       final val = (valueModel.currentValue as int);
       edit = info.editParameter.firstWhere((final element) {
-        final lower = element.raw["Min"] as int?;
-        final upper = element.raw["Max"] as int?;
+        final lower = element.raw["Min"] as double?;
+        final upper = element.raw["Max"] as double?;
         if (lower != null && upper != null) {
           return val >= lower && val < upper;
         }
